feat: Dockerize the socket server
