+++ docker-compose.yml
@@ -9,3 +9,14 @@ services:
     restart: always
     ports:
       - "5432:5432"
+
+  socket-server:
+    build:
+      context: .
+      dockerfile: Dockerfile.socket-server
+      args:
+        SOCKET_SERVER_SSL_CERT_PATH: "${SOCKET_SERVER_SSL_CERT_PATH:-socket-server/certificate-localhost.pem}"
+        SOCKET_SERVER_SSL_KEY_PATH: "${SOCKET_SERVER_SSL_KEY_PATH:-socket-server/key-localhost.pem}"
+    ports:
+      - "8080:8080"
+      - "8443:8443"
