+++ socket-server/token-validate.ts
@@ -1,0 +1,1 @@
+../util/token-validate.ts
