--- socket-server/token.ts
@@ -1,1 +1,0 @@
-../util/token.ts
